Improves system setup and suspend handling.
Streamlines the setup process by installing necessary plugins along with Remmina and also includes the evolution-ews package for better Exchange support.

Simplifies the suspend configuration by removing redundant steps and clarifying the instructions. This also refines the message to focus on essential information and includes instructions to log out and log back in for all settings to take effect.